+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.2:
+
+* Rework the poller so that the RequestFactory can be changed as long as the
+  WebPoller is not active. It no longer needs to be supplied when creating the
+  WebPoller.
+
 ## 0.1:
 
 * Initial release